+++ database/migrations/2022_06_25_075801_create_gostaresh_number_of_volunteers_status_analyses_table.php
@@ -30,8 +30,8 @@ return new class extends Migration
             $table->foreignId('minor_id')->nullable();
 
             $table->tinyInteger('university_type')->nullable();
-            $table->unsignedBigInteger('number_of_volunteers')->nullable();
-            $table->tinyInteger('department_of_education')->nullable(); 
+            $table->float('number_of_volunteers')->nullable();
+            $table->tinyInteger('department_of_education')->nullable();
 
             $table->integer('year')->unsigned()->nullable();
             $table->tinyInteger('month')->unsigned()->nullable();
