+++ database/migrations/2022_08_10_122837_create_gostaresh_number_of_students_status_by_grade_analyses_table.php
@@ -29,7 +29,7 @@ return new class extends Migration
             $table->foreignId('minor_id')->nullable();
 
             $table->tinyInteger('university_type')->nullable();
-            $table->unsignedBigInteger('number_of_students')->nullable();
+            $table->float('number_of_students')->nullable();
             $table->tinyInteger('department_of_education')->nullable();
 
             $table->integer('year')->unsigned()->nullable();
